chore: restructure integrations apps and entry points
