+++ src/integrations/FilesPublicShare/FilesPublicShareSidebar.vue
@@ -43,33 +43,33 @@ import { loadState } from '@nextcloud/initial-state'
 import { t } from '@nextcloud/l10n'
 import NcButton from '@nextcloud/vue/components/NcButton'
 import NcLoadingIcon from '@nextcloud/vue/components/NcLoadingIcon'
-import CallFailedDialog from './components/CallView/CallFailedDialog.vue'
-import CallView from './components/CallView/CallView.vue'
-import MediaSettings from './components/MediaSettings/MediaSettings.vue'
-import PollManager from './components/PollViewer/PollManager.vue'
-import PollViewer from './components/PollViewer/PollViewer.vue'
-import InternalSignalingHint from './components/RightSidebar/InternalSignalingHint.vue'
-import CallButton from './components/TopBar/CallButton.vue'
-import TopBar from './components/TopBar/TopBar.vue'
-import TransitionWrapper from './components/UIShared/TransitionWrapper.vue'
-import { useGetMessagesProvider } from './composables/useGetMessages.ts'
-import { useHashCheck } from './composables/useHashCheck.js'
-import { useIsInCall } from './composables/useIsInCall.js'
-import { useRecordingStatusSync } from './composables/useRecordingStatusSync.ts'
-import { useSessionIssueHandler } from './composables/useSessionIssueHandler.ts'
-import { EventBus } from './services/EventBus.ts'
-import { getPublicShareConversationData } from './services/filesIntegrationServices.ts'
+import CallFailedDialog from '../../components/CallView/CallFailedDialog.vue'
+import CallView from '../../components/CallView/CallView.vue'
+import MediaSettings from '../../components/MediaSettings/MediaSettings.vue'
+import PollManager from '../../components/PollViewer/PollManager.vue'
+import PollViewer from '../../components/PollViewer/PollViewer.vue'
+import InternalSignalingHint from '../../components/RightSidebar/InternalSignalingHint.vue'
+import CallButton from '../../components/TopBar/CallButton.vue'
+import TopBar from '../../components/TopBar/TopBar.vue'
+import TransitionWrapper from '../../components/UIShared/TransitionWrapper.vue'
+import { useGetMessagesProvider } from '../../composables/useGetMessages.ts'
+import { useHashCheck } from '../../composables/useHashCheck.js'
+import { useIsInCall } from '../../composables/useIsInCall.js'
+import { useRecordingStatusSync } from '../../composables/useRecordingStatusSync.ts'
+import { useSessionIssueHandler } from '../../composables/useSessionIssueHandler.ts'
+import { EventBus } from '../../services/EventBus.ts'
+import { getPublicShareConversationData } from '../../services/filesIntegrationServices.ts'
 import {
 	leaveConversationSync,
-} from './services/participantsService.js'
-import SessionStorage from './services/SessionStorage.js'
-import { useActorStore } from './stores/actor.ts'
-import { useTokenStore } from './stores/token.ts'
-import { checkBrowser } from './utils/browserCheck.ts'
-import { signalingKill } from './utils/webrtc/index.js'
+} from '../../services/participantsService.js'
+import SessionStorage from '../../services/SessionStorage.js'
+import { useActorStore } from '../../stores/actor.ts'
+import { useTokenStore } from '../../stores/token.ts'
+import { checkBrowser } from '../../utils/browserCheck.ts'
+import { signalingKill } from '../../utils/webrtc/index.js'
 
 export default {
-	name: 'PublicShareSidebar',
+	name: 'FilesPublicShareSidebar',
 
 	components: {
 		InternalSignalingHint,
