+++ src/integrations/FilesPublicShare/FilesPublicShareSidebarTrigger.vue
