+++ src/integrations/FilesSidebar/FilesSidebarChatView.vue
@@ -17,15 +17,15 @@
 
 <script>
 
-import CallFailedDialog from '../components/CallView/CallFailedDialog.vue'
-import MediaSettings from '../components/MediaSettings/MediaSettings.vue'
-import PollManager from '../components/PollViewer/PollManager.vue'
-import PollViewer from '../components/PollViewer/PollViewer.vue'
-import InternalSignalingHint from '../components/RightSidebar/InternalSignalingHint.vue'
-import CallButton from '../components/TopBar/CallButton.vue'
-import { useGetMessagesProvider } from '../composables/useGetMessages.ts'
-import { useGetToken } from '../composables/useGetToken.ts'
-import { useIsInCall } from '../composables/useIsInCall.js'
+import CallFailedDialog from '../../components/CallView/CallFailedDialog.vue'
+import MediaSettings from '../../components/MediaSettings/MediaSettings.vue'
+import PollManager from '../../components/PollViewer/PollManager.vue'
+import PollViewer from '../../components/PollViewer/PollViewer.vue'
+import InternalSignalingHint from '../../components/RightSidebar/InternalSignalingHint.vue'
+import CallButton from '../../components/TopBar/CallButton.vue'
+import { useGetMessagesProvider } from '../../composables/useGetMessages.ts'
+import { useGetToken } from '../../composables/useGetToken.ts'
+import { useIsInCall } from '../../composables/useIsInCall.js'
 
 export default {
