+++ src/integrations/FilesSidebar/FilesSidebarTabApp.vue
@@ -10,18 +10,18 @@
 <script>
 import { getCurrentUser } from '@nextcloud/auth'
 import { loadState } from '@nextcloud/initial-state'
-import FilesSidebarCallView from './views/FilesSidebarCallView.vue'
-import FilesSidebarChatView from './views/FilesSidebarChatView.vue'
-import { useIsInCall } from './composables/useIsInCall.js'
-import { useRecordingStatusSync } from './composables/useRecordingStatusSync.ts'
-import { useSessionIssueHandler } from './composables/useSessionIssueHandler.ts'
-import { EventBus } from './services/EventBus.ts'
-import { leaveConversationSync } from './services/participantsService.js'
-import SessionStorage from './services/SessionStorage.js'
-import { useActorStore } from './stores/actor.ts'
-import { useTokenStore } from './stores/token.ts'
-import { checkBrowser } from './utils/browserCheck.ts'
-import { signalingKill, signalingWebRtcKill } from './utils/webrtc/index.js'
+import FilesSidebarCallView from './FilesSidebarCallView.vue'
+import FilesSidebarChatView from './FilesSidebarChatView.vue'
+import { useIsInCall } from '../../composables/useIsInCall.js'
+import { useRecordingStatusSync } from '../../composables/useRecordingStatusSync.ts'
+import { useSessionIssueHandler } from '../../composables/useSessionIssueHandler.ts'
+import { EventBus } from '../../services/EventBus.ts'
+import { leaveConversationSync } from '../../services/participantsService.js'
+import SessionStorage from '../../services/SessionStorage.js'
+import { useActorStore } from '../../stores/actor.ts'
+import { useTokenStore } from '../../stores/token.ts'
+import { checkBrowser } from '../../utils/browserCheck.ts'
+import { signalingKill, signalingWebRtcKill } from '../../utils/webrtc/index.js'
 
 let fetchCurrentConversationIntervalId
