+++ src/integrations/deck.js
@@ -8,8 +8,8 @@ import { showError, showSuccess } from '@nextcloud/dialogs'
 import { t } from '@nextcloud/l10n'
 import { generateFilePath, generateUrl } from '@nextcloud/router'
 import escapeHtml from 'escape-html'
-import { postRichObjectToConversation } from './services/messagesService.ts'
-import { requestRoomSelection } from './utils/requestRoomSelection.js'
+import { postRichObjectToConversation } from '../services/messagesService.ts'
+import { requestRoomSelection } from '../utils/requestRoomSelection.js'
 
 import '@nextcloud/dialogs/style.css'
