+++ src/integrations/filesPublicShareSidebar.js
@@ -7,12 +7,12 @@ import { getCSPNonce } from '@nextcloud/auth'
 import { generateFilePath } from '@nextcloud/router'
 import { getSharingToken } from '@nextcloud/sharing/public'
 import { createApp, reactive } from 'vue'
-import PublicShareSidebar from './PublicShareSidebar.vue'
-import PublicShareSidebarTrigger from './PublicShareSidebarTrigger.vue'
-import { initializeTalkOnce } from './init.js'
-import { createMemoryRouter } from './router/router.ts'
-import store from './store/index.js'
-import pinia from './stores/pinia.ts'
+import FilesPublicShareSidebar from './FilesPublicShare/FilesPublicShareSidebar.vue'
+import FilesPublicShareSidebarTrigger from './FilesPublicShare/FilesPublicShareSidebarTrigger.vue'
+import { createMemoryRouter } from '../router/router.ts'
+import store from '../store/index.js'
+import pinia from '../stores/pinia.ts'
+import { initializeTalkOnce } from '../utils/init.js'
 
 initializeTalkOnce()
 
@@ -48,7 +48,7 @@ function addTalkSidebarTrigger() {
 	const mountPoint = document.querySelector('.header-end') ?? document.getElementById('header')
 	mountPoint.appendChild(talkSidebarTriggerElement)
 
-	createApp(PublicShareSidebarTrigger, {
+	createApp(FilesPublicShareSidebarTrigger, {
 		sidebarState,
 		onClick: () => {
 			sidebarState.isOpen = !sidebarState.isOpen
@@ -68,7 +68,7 @@ function addTalkSidebar() {
 
 	const router = createMemoryRouter()
 
-	createApp(PublicShareSidebar, {
+	createApp(FilesPublicShareSidebar, {
 		shareToken: getSharingToken(),
 		state: sidebarState,
 	})
