+++ src/integrations/search.js
@@ -7,10 +7,10 @@ import { getCSPNonce } from '@nextcloud/auth'
 import { emit } from '@nextcloud/event-bus'
 import { t } from '@nextcloud/l10n'
 import { generateFilePath, imagePath } from '@nextcloud/router'
-import { requestRoomSelection } from './utils/requestRoomSelection.js'
+import { requestRoomSelection } from '../utils/requestRoomSelection.js'
 
 import '@nextcloud/dialogs/style.css'
-import './assets/unified-search.css'
+import '../assets/unified-search.css'
 
 /**
  *
