+++ src/main.js
@@ -8,12 +8,12 @@ import { emit, subscribe } from '@nextcloud/event-bus'
 import { generateFilePath } from '@nextcloud/router'
 import { createApp, reactive, watch } from 'vue'
 import App from './App.vue'
-import { initializeTalkOnce } from './init.js'
 import { createTalkRouter } from './router/router.ts'
 import { SettingsAPI } from './services/SettingsAPI.ts'
 import store from './store/index.js'
 import pinia from './stores/pinia.ts'
 import { useSidebarStore } from './stores/sidebar.ts'
+import { initializeTalkOnce } from './utils/init.js'
 
 initializeTalkOnce()
